+++ src/main/java/com/example/library/domain/RentalDto.java
@@ -8,9 +8,9 @@ import java.util.Date;
 @Getter
 @AllArgsConstructor
 public class RentalDto {
-    private int id;
-    private int copyOfBookId;
-    private int readerId;
+    private Long id;
+    private Long copyOfBookId;
+    private Long readerId;
     private Date rentDate;
     private Date returnDate;
 }
